changed parameter type of validateValidation function to int64

+++ main.go
@@ -42,7 +42,7 @@ func main() {
 
 }
 
-func validateValidaiton(i float64){
+func validateValidaiton(i int64){
   for i <10{
     fmt.Println("Yeah!!!")
   }
